docs-content/02_rag/src/session1/requirements.txt
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.5
chromadb==0.4.22
tiktoken==0.5.2
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
numpy==1.24.3